refactor: TotalMilestoneScores 타입명을 MilestoneOverviewScore로 변경
