+++ frontend/src/types/milestone.ts
@@ -1,0 +1,6 @@
+export interface MilestoneOverviewScore {
+  activityScore: number;
+  globalScore: number;
+  communityScore: number;
+  totalScore: number;
+}
